requirements.txt
openai
langchain
os
python-dotenv
